requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0
asyncpg>=0.27.0
alembic>=1.11.0
python-jose[cryptography]>=3.1.0
passlib[bcrypt]>=1.7.0
redis>=5.0.0
httpx>=0.26.0
llama_cpp_python>=0.3.20
python-dotenv>=1.0.0
